refactor: orders wave 6 slices D–E — probe extract, connectivity fixtures (#4246)
Extract work-order tile-key probe preparation into a dedicated module and
densify connectivity dev target tests via shared fixtures; ratchet package
test LOC ceiling 16200 → 16095.
