+++ SPEC/program/repo-lint.md
@@ -149,7 +149,7 @@
 | `tool/check_orders_test_mirrors_lib.dart` | Forbid loose Dart sources directly at the `packages/colonizethis_orders/test/` root: orders production code lives under `lib/src/orders/`, so the testing rule (`colonizethis-testing.mdc`) requires `test/` to mirror `lib/`. Every orders test (and its co-located `*_test_support.dart` helper) must live in a mirroring subdirectory such as `test/orders/` (with `validators/`, `validators/diplomatic/`, `work_handlers/` subtrees) rather than flat in `test/`. A file is flagged when the path remainder after the `packages/colonizethis_orders/test/` prefix contains no further separator; nested files such as `test/orders/order_merge_part1_test.dart` are compliant. Mirrors the sibling combat layout gate `repo.combat_test_mirrors_lib` (Refs #3714); rule `repo.orders_test_mirrors_lib` |
 | `tool/check_orders_test_preserved_descriptions.dart` | Ensures every single-line `test('…')` / `testWidgets('…')` description in `packages/colonizethis_orders/test/DESCRIPTION_BASELINE.txt` still appears in the orders test tree (including `label:` pins under `test/orders/support/`) after wave-3 runner consolidation (Refs #3949); rule `repo.orders_test_preserved_descriptions` |
 | `tool/check_orders_test_support_layout.dart` | Non-`*_test.dart` Dart under `packages/colonizethis_orders/test/orders/` must live under `test/orders/support/` (wave-3 support consolidation; Refs #3949); rule `repo.orders_test_support_layout` |
-| `tool/check_orders_test_support_loc.dart` | Physical-line ceilings for `packages/colonizethis_orders/test/orders/support/**/*.dart` (wave-5 support AC ≤13,950) and package `test/**/*.dart` (wave-5 package AC ≤15,800; Refs #4109); rule `repo.orders_test_support_loc` |
+| `tool/check_orders_test_support_loc.dart` | Physical-line ceilings for `packages/colonizethis_orders/test/orders/support/**/*.dart` (wave-5 support AC ≤13,950) and package `test/**/*.dart` (wave-6 package AC ≤16,095 after connectivity dev fixture densify; Refs #4109, #4246); rule `repo.orders_test_support_loc` |
 | `tool/check_orders_test_shorthand_file_size.dart` | Physical-line ceiling (**180**) for every `*expectation_shorthand.dart` under `packages/colonizethis_orders/test/orders/support/**` (Refs #4109 wave 5); rule `repo.orders_test_shorthand_file_size` |
 | `tool/check_orders_lib_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_orders/lib/**` physical lines per file (wave-6 ≤350; empty shrink-only grandfather after Slice A; generated suffixes excluded); rule `repo.orders_lib_source_file_size` (Refs #4246) |
 | `tool/check_app_flame_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/flame/**` physical lines per file (wave-10 ≤400; shrink-only grandfather allowlist for region-map render modules and near-cap province-action states during transition; generated suffixes excluded); rule `repo.app_flame_lib_physical_file_size` (Refs #4183) |
