+++ packages/colonizethis_orders/test/orders/connectivity_dev_targets_test.dart
@@ -1,39 +1,17 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_orders/src/orders/connectivity_dev_snapshot.dart';
 import 'package:colonizethis_orders/src/orders/connectivity_dev_targets.dart';
 import 'package:colonizethis_orders/src/orders/feedstock_extraction_targets.dart';
 import 'package:colonizethis_orders/src/orders/order_work_constants.dart';
 import 'package:colonizethis_test/test.dart';
 
 import 'support/scenario_runner.dart';
+import 'support/suggestion/connectivity_dev_targets_fixtures.dart';
 import 'support/suggestion/order_suggestion_work_feedstock_priority_fixtures.dart';
-
-ConnectivityDevSnapshot _snapshot({
-  Set<String> connected = const {},
-  Set<String> frontier = const {},
-  Set<String> adjacent = const {},
-  Set<String> bottleneck = const {},
-  Map<String, int> extensionDistance = const {},
-  bool hasTargets = true,
-}) {
-  return ConnectivityDevSnapshot(
-    connected: connected,
-    pathTransportCap: const {},
-    extensionDistanceByTile: extensionDistance,
-    seaZonesReachableFromCapital: const {},
-    provincesWithUnconnectedDevTargets: const {},
-    hasUnconnectedDevTargets: hasTargets,
-    frontierExtensionTiles: frontier,
-    bottleneckRailTiles: bottleneck,
-    adjacentToConnectedTiles: adjacent,
-  );
-}
 
 void main() {
   runLabeledScenarioGroup('prioritizeBuildRoadCandidatesByConnectivity', [
     rs('AC-A2 frontier hard demotion', () {
-      final snapshot = _snapshot(
+      final snapshot = connectivityDevSnapshot(
         frontier: {'oldWorld|p1|0|1'},
         extensionDistance: {'oldWorld|p1|0|1': 2},
       );
@@ -44,7 +22,7 @@ void main() {
       expect(ordered.first, 'oldWorld|p1|0|1');
     }),
     rs('AC-A3 nearest-target-first within frontier', () {
-      final snapshot = _snapshot(
+      final snapshot = connectivityDevSnapshot(
         frontier: {'oldWorld|p1|0|1', 'oldWorld|p1|0|3'},
         extensionDistance: {'oldWorld|p1|0|1': 1, 'oldWorld|p1|0|3': 3},
       );
@@ -55,7 +33,7 @@ void main() {
       expect(ordered, ['oldWorld|p1|0|1', 'oldWorld|p1|0|3']);
     }),
     rs('AC-A4 baseline when no unconnected targets', () {
-      final snapshot = _snapshot(hasTargets: false);
+      final snapshot = connectivityDevSnapshot(hasTargets: false);
       const input = ['oldWorld|p1|9|9', 'oldWorld|p1|0|1'];
       expect(
         prioritizeBuildRoadCandidatesByConnectivity(
@@ -69,7 +47,7 @@ void main() {
 
   runLabeledScenarioGroup('prioritizeBuildImprovementCandidatesByConnectivity', [
     rs('AC-C1 connected > adjacent > far', () {
-      final snapshot = _snapshot(
+      final snapshot = connectivityDevSnapshot(
         connected: {'c'},
         adjacent: {'a'},
       );
@@ -93,9 +71,8 @@ void main() {
           ),
           contains('iron'),
         );
-        final snapshot = _snapshot(
+        final snapshot = connectivityDevSnapshot(
           connected: {feedstockPrioritySupplierGrainTile},
-          adjacent: const {},
         );
         final ordered = applyBuildImprovementConnectivityPreservingFeedstock(
           game: game,
@@ -114,7 +91,7 @@ void main() {
 
   runLabeledScenarioGroup('prioritizeBuildRailCandidatesByConnectivity', [
     rs('AC-B1 bottleneck promotion', () {
-      final snapshot = _snapshot(
+      final snapshot = connectivityDevSnapshot(
         connected: {'bottleneck', 'plain'},
         bottleneck: {'bottleneck'},
       );
@@ -125,7 +102,7 @@ void main() {
       expect(ordered.first, 'bottleneck');
     }),
     rs('AC-B2 no-yield-gain demotion', () {
-      final snapshot = _snapshot(
+      final snapshot = connectivityDevSnapshot(
         connected: {'bottleneck', 'satisfied'},
         bottleneck: {'bottleneck'},
       );
@@ -139,178 +116,28 @@ void main() {
 
   runLabeledScenarioGroup('prioritizeBuildPortCandidatesByConnectivity', [
     rs('AC-D1 overseas resource province promotion', () {
-      const ow = 'oldWorld';
-      const provinceId = '$ow|nw1';
-      const linkedPort = '$provinceId|0|0';
-      const plainPort = 'oldWorld|ow1|0|0';
-      final snapshot = ConnectivityDevSnapshot(
-        connected: const {},
-        pathTransportCap: const {},
-        extensionDistanceByTile: const {},
-        seaZonesReachableFromCapital: {'$ow|sea1'},
-        provincesWithUnconnectedDevTargets: {provinceId},
-        hasUnconnectedDevTargets: true,
-        frontierExtensionTiles: const {},
-        bottleneckRailTiles: const {},
-        adjacentToConnectedTiles: const {},
+      final c = connectivityDevOverseasResourcePortCase();
+      expect(
+        prioritizeBuildPortCandidatesByConnectivity(
+          snapshot: c.snapshot,
+          sortedVisible: c.visible,
+          game: c.game,
+          topology: c.topology,
+        ).first,
+        c.expectedFirst,
       );
-      final game = Game(
-        id: 'g',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: [
-              Province(id: 'oldWorld|ow1', regionId: ow, ownerId: 'gp1'),
-              Province(id: provinceId, regionId: ow, ownerId: 'gp1'),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: [
-          Player(
-            id: 'gp1',
-            displayName: 'GP',
-            isHuman: false,
-            capitalProvinceId: 'oldWorld|ow1',
-            capitalTile: CapitalTile(
-              regionId: ow,
-              provinceId: 'oldWorld|ow1',
-              x: 0,
-              y: 0,
-            ),
-          ),
-        ],
-      );
-      const topology = MapTopology(
-        nodes: [
-          TopologyNode(
-            id: 'oldWorld|ow1',
-            regionId: ow,
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: provinceId,
-            regionId: ow,
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: '$ow|sea1',
-            regionId: ow,
-            type: TopologyNodeType.seaZone,
-          ),
-        ],
-        edges: [
-          TopologyEdge(
-            id1: 'oldWorld|ow1',
-            id2: '$ow|sea1',
-          ),
-          TopologyEdge(
-            id1: provinceId,
-            id2: '$ow|sea1',
-          ),
-        ],
-      );
-      final ordered = prioritizeBuildPortCandidatesByConnectivity(
-        snapshot: snapshot,
-        sortedVisible: [plainPort, linkedPort],
-        game: game,
-        topology: topology,
-      );
-      expect(ordered.first, linkedPort);
     }),
     rs('AC-D2 sea-unreachable demotion', () {
-      const ow = 'oldWorld';
-      const reachableProvince = '$ow|nwReach';
-      const unreachableProvince = '$ow|nwBlock';
-      const reachablePort = '$reachableProvince|0|0';
-      const unreachablePort = '$unreachableProvince|0|0';
-      final snapshot = ConnectivityDevSnapshot(
-        connected: const {},
-        pathTransportCap: const {},
-        extensionDistanceByTile: const {},
-        seaZonesReachableFromCapital: {'$ow|seaReach'},
-        provincesWithUnconnectedDevTargets: {
-          reachableProvince,
-          unreachableProvince,
-        },
-        hasUnconnectedDevTargets: true,
-        frontierExtensionTiles: const {},
-        bottleneckRailTiles: const {},
-        adjacentToConnectedTiles: const {},
+      final c = connectivityDevSeaUnreachablePortCase();
+      expect(
+        prioritizeBuildPortCandidatesByConnectivity(
+          snapshot: c.snapshot,
+          sortedVisible: c.visible,
+          game: c.game,
+          topology: c.topology,
+        ).first,
+        c.expectedFirst,
       );
-      final game = Game(
-        id: 'g',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: [
-              Province(
-                id: 'oldWorld|ow1',
-                regionId: ow,
-                ownerId: 'gp1',
-              ),
-              Province(id: reachableProvince, regionId: ow, ownerId: 'gp1'),
-              Province(id: unreachableProvince, regionId: ow, ownerId: 'gp1'),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: [
-          Player(
-            id: 'gp1',
-            displayName: 'GP',
-            isHuman: false,
-            capitalProvinceId: 'oldWorld|ow1',
-            capitalTile: CapitalTile(
-              regionId: ow,
-              provinceId: 'oldWorld|ow1',
-              x: 0,
-              y: 0,
-            ),
-          ),
-        ],
-      );
-      const topology = MapTopology(
-        nodes: [
-          TopologyNode(
-            id: 'oldWorld|ow1',
-            regionId: ow,
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: reachableProvince,
-            regionId: ow,
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: unreachableProvince,
-            regionId: ow,
-            type: TopologyNodeType.province,
-          ),
-          TopologyNode(
-            id: '$ow|seaReach',
-            regionId: ow,
-            type: TopologyNodeType.seaZone,
-          ),
-          TopologyNode(
-            id: '$ow|seaBlock',
-            regionId: ow,
-            type: TopologyNodeType.seaZone,
-          ),
-        ],
-        edges: [
-          TopologyEdge(id1: 'oldWorld|ow1', id2: '$ow|seaReach'),
-          TopologyEdge(id1: reachableProvince, id2: '$ow|seaReach'),
-          TopologyEdge(id1: unreachableProvince, id2: '$ow|seaBlock'),
-        ],
-      );
-      final ordered = prioritizeBuildPortCandidatesByConnectivity(
-        snapshot: snapshot,
-        sortedVisible: [unreachablePort, reachablePort],
-        game: game,
-        topology: topology,
-      );
-      expect(ordered.first, reachablePort);
     }),
   ], runRunnableScenario);
 
@@ -326,22 +153,13 @@ void main() {
 
   runLabeledScenarioGroup('applyConnectivityDevTargetOrdering', [
     rs('unknown target unchanged', () {
-      final snapshot = _snapshot();
       const input = ['t1', 't2'];
       expect(
         applyConnectivityDevTargetOrdering(
           workTarget: kWorkTargetExplore,
           sortedVisible: input,
-          snapshot: snapshot,
-          game: Game(
-            id: 'g',
-            worldState: WorldState(
-              turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-              oldWorld: const RegionData(),
-              newWorld: const RegionData(),
-            ),
-            players: const [],
-          ),
+          snapshot: connectivityDevSnapshot(),
+          game: connectivityDevEmptyGame(),
           topology: const MapTopology(nodes: [], edges: []),
           tileMapByRegion: const {},
         ),
