+++ tool/check_orders_test_support_loc.dart
@@ -20,9 +20,8 @@ const String ordersTestPackageRelativeDir = 'packages/colonizethis_orders/test';
 const int ordersTestSupportLocCeiling = 13950;
 
 /// Ratchet ceiling for package `test/` physical LOC.
-/// Wave-5 AC target: 15800; raised for naval mission draft/availability suite (#4213),
-/// then Spy relocate projected-tile scenarios (#4219).
-const int ordersTestPackageLocCeiling = 16200;
+/// Wave-6 Slice E (#4246): connectivity dev fixtures densify 16200 → 16095.
+const int ordersTestPackageLocCeiling = 16095;
 
 /// Counts physical lines of all `*.dart` files under [dir].
 int countOrdersTestSupportPhysicalLoc(Directory dir) {
@@ -94,13 +93,13 @@ int runCheckOrdersTestSupportLoc(
   if (packageLoc > packageCeiling) {
     logE(
       'check_orders_test_support_loc: package test/ LOC $packageLoc exceeds '
-      'ceiling $packageCeiling (wave-5 target ≤16200; Refs #4109, #4213, #4219).',
+      'ceiling $packageCeiling (wave-6 target ≤16095; Refs #4246).',
     );
     return 1;
   }
   logI(
     'check_orders_test_support_loc: package test/ LOC $packageLoc ≤ ceiling '
-    '$packageCeiling (wave-5 target ≤16200; Refs #4109, #4213, #4219).',
+    '$packageCeiling (wave-6 target ≤16095; Refs #4246).',
   );
   return 0;
 }
